move: "client/~/assets.js를 init에서 utils로 이동
- utils 폴더 생성
- 클라이언트측 assets.js는 초기화 용도로 사용되지 않기 때문에 목적성을 고려하여 utils로 이동
